Code refactoring

+++ location/processors/Location.Processors/Subscribers/BookingSubscriber.cs
@@ -12,10 +12,7 @@ using Type = Api.Shared.Clients.Events.Skedular.Booking.V1.Value.Type;
 
 namespace Location.Processors.Subscribers;
 
-public class BookingSubscriber(
-    ILogger<BookingSubscriber> logger,
-    IMapper mapper,
-    IRepositoryFactory repositoryFactory) : IEventSubscriber<Key, Event>
+public class BookingSubscriber(ILogger<BookingSubscriber> logger, IMapper mapper, IRepositoryFactory repositoryFactory) : IEventSubscriber<Key, Event>
 {
     public async Task<EventSubscriberResult> HandleAsync(EventContext eventContext, Key key, Event @event, CancellationToken cancellationToken)
     {
@@ -24,35 +21,29 @@ public class BookingSubscriber(
             case Type.BookingUpserted:
                 {
                     var booking = mapper.MapTo(@event);
-                    var existingBooking = await repositoryFactory.BookingRepository.GetByIdAsync(booking.Id, cancellationToken);
-                    if (existingBooking is not null && existingBooking.EventRaisedAt > booking.EventRaisedAt)
+                    if (string.IsNullOrWhiteSpace(booking.Location.Id))
                     {
-                        logger.LogInformation("Ignoring Booking event. Event timestamp is older that what is already processed.");
-
-                        return EventSubscriberResults.Success;
+                        await HandleBookingDeletedEventAsync(booking, cancellationToken);
                     }
+                    else
+                    {
+                        var existingBooking = await repositoryFactory.BookingRepository.UpsertNakedAsync(booking.Id, cancellationToken);
+                        if (existingBooking.EventRaisedAt > booking.EventRaisedAt)
+                        {
+                            logger.LogInformation("Ignoring Booking event. Event timestamp is older that what is already processed.");
 
-                    await HandleBookingUpsertedEventAsync(booking, existingBooking, cancellationToken);
+                            return EventSubscriberResults.Success;
+                        }
+
+                        await HandleBookingUpsertedEventAsync(booking, existingBooking, cancellationToken);
+                    }
                 }
                 break;
 
             case Type.BookingDeleted:
                 {
                     var booking = mapper.MapTo(@event);
-                    var existingBooking = await repositoryFactory.BookingRepository.GetByIdAsync(booking.Id, cancellationToken);
-                    if (existingBooking is not null && existingBooking.EventRaisedAt > booking.EventRaisedAt)
-                    {
-                        logger.LogInformation("Ignoring Booking event. Event timestamp is older that what is already processed.");
-
-                        return EventSubscriberResults.Success;
-                    }
-
-                    if (existingBooking is null)
-                    {
-                        return EventSubscriberResults.Success;
-                    }
-
-                    await HandleBookingDeletedEventAsync(existingBooking, cancellationToken);
+                    await HandleBookingDeletedEventAsync(booking, cancellationToken);
                 }
                 break;
         }
@@ -60,23 +51,8 @@ public class BookingSubscriber(
         return EventSubscriberResults.Success;
     }
 
-    private async Task HandleBookingUpsertedEventAsync(Shared.Models.Booking booking, Booking? existingBooking, CancellationToken cancellationToken)
+    private async Task HandleBookingUpsertedEventAsync(Shared.Models.Booking booking, Booking existingBooking, CancellationToken cancellationToken)
     {
-        if (existingBooking is not null && string.IsNullOrWhiteSpace(booking.Location.Id))
-        {
-            // If booking already exist and is now detached from location, delete it
-            _ = repositoryFactory.BookingRepository.Remove(existingBooking);
-            await repositoryFactory.UnitOfWork.SaveChangesAsync(cancellationToken);
-
-            return;
-        }
-
-        if (string.IsNullOrWhiteSpace(booking.Location.Id))
-        {
-            // Booking not attached to any location, ignoring it
-            return;
-        }
-
         var resources = new List<Resource>();
         if (booking.Resources.Count != 0)
         {
@@ -93,15 +69,26 @@ public class BookingSubscriber(
             throw new LocationNotFound();
         }
 
-        _ = existingBooking is null
-            ? repositoryFactory.BookingRepository.Add(mapper.MapToEntity(booking, location, resources))
-            : repositoryFactory.BookingRepository.Update(mapper.MergeToEntity(booking, existingBooking, location, resources));
+        _ = repositoryFactory.BookingRepository.Update(mapper.MergeToEntity(booking, existingBooking, location, resources));
 
         await repositoryFactory.UnitOfWork.SaveChangesAsync(cancellationToken);
     }
 
-    private async Task HandleBookingDeletedEventAsync(Booking existingBooking, CancellationToken cancellationToken)
+    private async Task HandleBookingDeletedEventAsync(Shared.Models.Booking booking, CancellationToken cancellationToken)
     {
+        var existingBooking = await repositoryFactory.BookingRepository.GetByIdAsync(booking.Id, cancellationToken);
+        if (existingBooking is not null && existingBooking.EventRaisedAt > booking.EventRaisedAt)
+        {
+            logger.LogInformation("Ignoring Booking event. Event timestamp is older that what is already processed.");
+
+            return;
+        }
+
+        if (existingBooking is null)
+        {
+            return;
+        }
+
         existingBooking.Resources = [];
         existingBooking = repositoryFactory.BookingRepository.Update(existingBooking);
         _ = repositoryFactory.BookingRepository.Remove(existingBooking);
