+++ organization/processors/Organization.Processors/Subscribers/LocationSubscriber.cs
@@ -1,4 +1,5 @@
 using Api.Shared.Clients.Events.Skedular.Location.V1.Key;
+using Enterprise.Shared.Exceptions;
 using Enterprise.Shared.Kafka.Consume;
 using Organization.Processors.Mappers;
 using Organization.Shared.Repositories;
@@ -8,10 +9,8 @@ using Type = Api.Shared.Clients.Events.Skedular.Location.V1.Value.Type;
 
 namespace Organization.Processors.Subscribers;
 
-public class LocationSubscriber(
-    ILogger<LocationSubscriber> logger,
-    IMapper mapper,
-    IRepositoryFactory repositoryFactory) : IEventSubscriber<Key, Event>
+public class LocationSubscriber(ILogger<LocationSubscriber> logger, IMapper mapper, IRepositoryFactory repositoryFactory)
+    : IEventSubscriber<Key, Event>
 {
     public async Task<EventSubscriberResult> HandleAsync(EventContext eventContext, Key key, Event @event, CancellationToken cancellationToken)
     {
@@ -20,49 +19,36 @@ public class LocationSubscriber(
             case Type.LocationUpserted:
                 {
                     var location = mapper.MapTo(@event);
-                    if (string.IsNullOrWhiteSpace(@event.Data.Location.OrganizationId))
+                    if (string.IsNullOrWhiteSpace(location.Organization.Id))
                     {
-                        break;
+                        await HandleLocationDeletedEventAsync(location, cancellationToken);
                     }
-
-                    var existingOrganization =
-                        await repositoryFactory.OrganizationRepository.GetByIdAsync(location.Organization.Id, cancellationToken);
-                    ArgumentNullException.ThrowIfNull(existingOrganization);
-
-                    var existingLocation =
-                        await repositoryFactory.LocationRepository.UpsertNakedAsync(location.Id, existingOrganization, cancellationToken);
-                    if (existingLocation.EventRaisedAt > location.EventRaisedAt)
+                    else
                     {
-                        logger.LogInformation("Ignoring Location event. Event timestamp is older that what is already processed.");
+                        var organization = await repositoryFactory.OrganizationRepository.GetByIdAsync(location.Organization.Id, cancellationToken);
+                        if (organization is null)
+                        {
+                            throw new OrganizationNotFound();
+                        }
 
-                        return EventSubscriberResults.Success;
+                        var existingLocation =
+                            await repositoryFactory.LocationRepository.UpsertNakedAsync(location.Id, organization, cancellationToken);
+                        if (existingLocation.EventRaisedAt > location.EventRaisedAt)
+                        {
+                            logger.LogInformation("Ignoring Location event. Event timestamp is older that what is already processed.");
+
+                            return EventSubscriberResults.Success;
+                        }
+
+                        await HandleLocationUpsertedEventAsync(location, existingLocation, organization, cancellationToken);
                     }
-
-                    await HandleLocationUpsertedEventAsync(
-                        location,
-                        existingLocation,
-                        existingOrganization,
-                        cancellationToken);
                 }
                 break;
 
             case Type.LocationDeleted:
                 {
                     var location = mapper.MapTo(@event);
-                    var existingLocation = await repositoryFactory.LocationRepository.GetByIdAsync(location.Id, cancellationToken);
-                    if (existingLocation is not null && existingLocation.EventRaisedAt > location.EventRaisedAt)
-                    {
-                        logger.LogInformation("Ignoring Location event. Event timestamp is older that what is already processed.");
-
-                        return EventSubscriberResults.Success;
-                    }
-
-                    if (existingLocation is null)
-                    {
-                        return EventSubscriberResults.Success;
-                    }
-
-                    await HandleLocationDeletedEventAsync(existingLocation, cancellationToken);
+                    await HandleLocationDeletedEventAsync(location, cancellationToken);
                 }
                 break;
         }
@@ -76,27 +62,25 @@ public class LocationSubscriber(
         Shared.Database.Entities.Organization existingOrganization,
         CancellationToken cancellationToken)
     {
-        if (string.IsNullOrWhiteSpace(location.Organization.Id))
-        {
-            // If location already exist and is now detached from organization, delete it
-            _ = repositoryFactory.LocationRepository.Remove(existingLocation);
-            await repositoryFactory.UnitOfWork.SaveChangesAsync(cancellationToken);
-
-            return;
-        }
-
-        if (string.IsNullOrWhiteSpace(location.Organization.Id))
-        {
-            // Location not attached to any organization, ignoring it
-            return;
-        }
-
         _ = repositoryFactory.LocationRepository.Update(mapper.MergeToEntity(location, existingLocation, existingOrganization));
         await repositoryFactory.UnitOfWork.SaveChangesAsync(cancellationToken);
     }
 
-    private async Task HandleLocationDeletedEventAsync(Location existingLocation, CancellationToken cancellationToken)
+    private async Task HandleLocationDeletedEventAsync(Shared.Models.Location location, CancellationToken cancellationToken)
     {
+        var existingLocation = await repositoryFactory.LocationRepository.GetByIdAsync(location.Id, cancellationToken);
+        if (existingLocation is not null && existingLocation.EventRaisedAt > location.EventRaisedAt)
+        {
+            logger.LogInformation("Ignoring Location event. Event timestamp is older that what is already processed.");
+
+            return;
+        }
+
+        if (existingLocation is null)
+        {
+            return;
+        }
+
         _ = repositoryFactory.LocationRepository.Remove(existingLocation);
         await repositoryFactory.UnitOfWork.SaveChangesAsync(cancellationToken);
     }
